refactor: StorageController.java is better defined as a service
