+++ src/main/java/com/om1cael/controller/MovieStorageController.java
@@ -1,21 +1,22 @@
 package com.om1cael.controller;
 
 import com.om1cael.model.Movie;
+import com.om1cael.services.StorageService;
 
 import java.util.List;
 
 public class MovieStorageController {
-    private final StorageController storageController;
+    private final StorageService storageService;
 
     public MovieStorageController() {
-        this.storageController = new StorageController();
+        this.storageService = new StorageService();
     }
 
     public void addToMovies(Movie movie) {
-        this.storageController.saveToFile(movie);
+        this.storageService.saveToFile(movie);
     }
 
     public List<Movie> getMovies() {
-        return this.storageController.readFromFile();
+        return this.storageService.readFromFile();
     }
 }
